fixed buy/sell & orderbook amounts and precision
Note: the limit order subscription notifications are still not
accurate, nor is the Recent Activity panel. However, the fill order
activity entries have accurate amounts. This leads me to conclude that
the orders are being placed correctly, but that the subscription
parsing and activity panel both need to be adjusted. (I could be wrong.)