SendModal - Close once 'Send' is clicked.
Per github discussion, due to the async nature of these modals it's much
simpler to just close to Send dialog and have a user re-open the Send
box if they have an error.

Included basic check to disable Send button from activating if the from
and to accounts are the same.