Creating a Code of Conduct
Adopting and following a code of conduct would be in the projects best interest. As a Project Manager I think this is a requirement to set a good base for what is not allowed and what our goal is. As no convinent e-mail address can be given, contact way is set to Telegram.